Move spdx/formats.py to spdxtagvalue directory
In preparation of creating a new SPDX JSON format, move spdx/formats.py to
the spdx/spdxtagvalue subdirectory, as the majority of the functions and
variables defined in this file are specific to the tag-value
formatting style.

Works towards #839
