+++ tern/formats/spdx/spdxtagvalue/package_helpers.py
@@ -7,7 +7,7 @@
 Helper functions for packages in SPDX document
 """
 
-from tern.formats.spdx import formats as spdx_formats
+from tern.formats.spdx.spdxtagvalue import formats as spdx_formats
 from tern.report import content
